+++ src/components/ActionButton/ActionButton.tsx
@@ -4,6 +4,11 @@ import * as React from 'react';
 import { getClassNames } from './ActionButton.classNames';
 import { ActionButtonProps } from './ActionButton.types';
 
+/**
+ * @deprecated Komponenten er erstattet av InlineButton fra "@skatteetaten/ds-buttons"
+ *
+ * visibleName ActionButton (Aksjonsknapp)
+ */
 export class ActionButton extends React.PureComponent<ActionButtonProps, {}> {
   static NORMAL = 'large';
   static LARGE = 'xlarge';