+++ src/components/Button/Button.tsx
@@ -4,10 +4,11 @@ import * as React from 'react';
 import { getClassNames as getStandardClassNames } from './Button.classNames';
 import { ButtonProps } from './Button.types';
 
-/*
+/**
+ * @deprecated Komponenten er erstattet av Button fra "@skatteetaten/ds-buttons"
+ *
  * visibleName Button (Knapp)
  */
-
 export const Button: React.FC<ButtonProps> = (props) => {
   const { children, icon, className, iconProps, buttonType, ...rest } = props;