+++ src/components/Icon/Icon.tsx
@@ -1,7 +1,9 @@
 import { Icon as FabricIcon, IIconProps } from '@fluentui/react';
 import * as React from 'react';
 
-/*
+/**
+ * @deprecated Komponenten er erstattet av Icon fra "@skatteetaten/ds-icons"
+ *
  * visibleName Icon (Ikon)
  */
 export const Icon: React.FC<IIconProps> = (props) => {