+++ src/components/IconButton/IconButton.tsx
@@ -4,8 +4,10 @@ import { IconButton as FabricIconButton } from '@fluentui/react';
 import { getClassNames } from './IconButton.classNames';
 import { IconButtonProps } from './IconButton.types';
 
-/*
- * visibleName IconButton (Ikonknapp)
+/**
+ * @deprecated Komponenten er erstattet av IconButton fra "@skatteetaten/ds-buttons"
+ *
+ * visibleName IconButton (IkonKnapp)
  */
 export const IconButton: React.FC<IconButtonProps> = (props) => {
   const { icon, className, ...rest } = props;