+++ src/migrations/index.tsx
@@ -1,0 +1,17 @@
+/* NB! Dette er en autogenerert fil, vennligst ikke gjør manuelle endringer i den.
+   Kjør kommandonavn generate:migrationGuidesIndexFile for å generere exports. */
+import ActionButtonMigration from './ActionButtonMigration.md';
+import ButtonLinkMigration from './ButtonLinkMigration.md';
+import ButtonMigration from './ButtonMigration.md';
+import IconButtonMigration from './IconButtonMigration.md';
+import IconMigration from './IconMigration.md';
+import SkeBasisMigration from './SkeBasisMigration.md';
+
+export const MigrationGuides = [
+  ActionButtonMigration,
+  ButtonLinkMigration,
+  ButtonMigration,
+  IconButtonMigration,
+  IconMigration,
+  SkeBasisMigration,
+];