+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+### **v.6.0.0 - November 2022**
+
+The package '@skatteetaten/frontend-components@6.0.0' is released in parallel with '@skatteetaten/ds-core-designtokens', '@skatteetaten/ds-buttons' and '@skatteetaten/ds-icons'. The components in these packages replace the ones marked 'deprecated' in the documentation. From now on this package (@skatteetaten/frontend-components/) will be called 'Designsystem legacy'.
+
+- Upgraded dependencies: @fluentui/react, i18next, react-i18next, typescript
+- Add children prop explicitly for improved compatibility with React 18
+- Moment: removal of moment as a dependency
+- Deprecating the following components:
+  - Designtokens and SkeBasis
+  - Button, ActionButton, IconButton and ButtonLink
+  - Icon
+
 ### **v.5.1.0 - 8 September 2022**
 
 - Dialog: new variant with waiting dialog.
@@ -17,7 +29,7 @@
 
 - SearchField: fixes an issue where the search button was not clickable when focused.
 - FileUploader: fixes an issue where aria-describedby was not set correctly with no accepted file formats.
-- Documentation: Fixes an issue with the path where "frontend-components" was missing. Updated examples: Dialog, Modal, RadioButtonGroup, Table (editable rows), and error messages. Updated typography section.
+- Documentation: fixes an issue with the path where "frontend-components" was missing. Updated examples: Dialog, Modal, RadioButtonGroup, Table (editable rows), and error messages. Updated typography section.
 
 ### **v.5.0.4 - 6 April 2022**